@@ -19,7 +19,7 @@ import shapely.ops
 import shapely.wkt
 from shapely.geometry.base import BaseGeometry
 
-from ._compat import HAS_PYPROJ, requires_pyproj
+from ._compat import HAS_PYPROJ, PANDAS_GE_21, PANDAS_GE_22, requires_pyproj
 from .sindex import SpatialIndex
 
 if HAS_PYPROJ:
@@ -58,14 +58,12 @@ class GeometryDtype(ExtensionDtype):
     def construct_from_string(cls, string):
         if not isinstance(string, str):
             raise TypeError(
-                "'construct_from_string' expects a string, got {}".format(type(string))
+                f"'construct_from_string' expects a string, got {type(string)}"
             )
         elif string == cls.name:
             return cls()
         else:
-            raise TypeError(
-                "Cannot construct a '{}' from '{}'".format(cls.__name__, string)
-            )
+            raise TypeError(f"Cannot construct a '{cls.__name__}' from '{string}'")
 
     @classmethod
     def construct_array_type(cls):
@@ -112,8 +110,8 @@ def _crs_mismatch_warn(left, right, stacklevel=3):
         "and the CRS of right geometries.\n"
         "Use `to_crs()` to reproject one of "
         "the input geometries to match the CRS of the other.\n\n"
-        "Left CRS: {0}\n"
-        "Right CRS: {1}\n".format(left_srs, right_srs),
+        f"Left CRS: {left_srs}\n"
+        f"Right CRS: {right_srs}\n",
         UserWarning,
         stacklevel=stacklevel,
     )
@@ -164,6 +162,8 @@ def from_shapely(data, crs=None):
     if not isinstance(data, np.ndarray):
         arr = np.empty(len(data), dtype=object)
         arr[:] = data
+    elif len(data) == 0 and data.dtype == "float64":
+        arr = data.astype(object)
     else:
         arr = data
 
@@ -179,9 +179,7 @@ def from_shapely(data, crs=None):
             elif isna(geom):
                 out.append(None)
             else:
-                raise TypeError(
-                    "Input must be valid geometry objects: {0}".format(geom)
-                )
+                raise TypeError(f"Input must be valid geometry objects: {geom}")
         arr = np.array(out, dtype=object)
 
     return GeometryArray(arr, crs=crs)
@@ -404,11 +402,10 @@ class GeometryArray(ExtensionArray):
         """Check CRS and warn if the planar operation is done in a geographic CRS"""
         if self.crs and self.crs.is_geographic:
             warnings.warn(
-                "Geometry is in a geographic CRS. Results from '{}' are likely "
-                "incorrect. Use 'GeoSeries.to_crs()' to re-project geometries to a "
-                "projected CRS before this operation.\n".format(
-                    inspect.stack()[1].function
-                ),
+                "Geometry is in a geographic CRS. Results from "
+                f"'{inspect.stack()[1].function}' are likely incorrect. "
+                "Use 'GeoSeries.to_crs()' to re-project geometries to a "
+                "projected CRS before this operation.\n",
                 UserWarning,
                 stacklevel=stacklevel,
             )
@@ -437,7 +434,7 @@ class GeometryArray(ExtensionArray):
             value = value.values
         if isinstance(value, pd.DataFrame):
             value = value.values.flatten()
-        if isinstance(value, (list, np.ndarray)):
+        if isinstance(value, list | np.ndarray):
             value = from_shapely(value)
         if isinstance(value, GeometryArray):
             if isinstance(key, numbers.Integral):
@@ -452,7 +449,7 @@ class GeometryArray(ExtensionArray):
                 value = from_shapely([value])._data[0]
             else:
                 raise TypeError("should be valid geometry")
-            if isinstance(key, (slice, list, np.ndarray)):
+            if isinstance(key, slice | list | np.ndarray):
                 value_array = np.empty(1, dtype=object)
                 value_array[:] = [value]
                 self._data[key] = value_array
@@ -460,7 +457,7 @@ class GeometryArray(ExtensionArray):
                 self._data[key] = value
         else:
             raise TypeError(
-                "Value should be either a BaseGeometry or None, got %s" % str(value)
+                f"Value should be either a BaseGeometry or None, got {value!s}"
             )
 
         # invalidate spatial index
@@ -691,8 +688,9 @@ class GeometryArray(ExtensionArray):
     def _binary_method(op, left, right, **kwargs):
         if isinstance(right, GeometryArray):
             if len(left) != len(right):
-                msg = "Lengths of inputs do not match. Left: {0}, Right: {1}".format(
-                    len(left), len(right)
+                msg = (
+                    "Lengths of inputs do not match. "
+                    f"Left: {len(left)}, Right: {len(right)}"
                 )
                 raise ValueError(msg)
             if not _check_crs(left, right):
@@ -809,7 +807,7 @@ class GeometryArray(ExtensionArray):
         return self._binary_method("frechet_distance", self, other, **kwargs)
 
     def buffer(self, distance, resolution=16, **kwargs):
-        if not (isinstance(distance, (int, float)) and distance == 0):
+        if not (isinstance(distance, int | float) and distance == 0):
             self.check_geographic_crs(stacklevel=5)
         return GeometryArray(
             shapely.buffer(self._data, distance, quad_segs=resolution, **kwargs),
@@ -1220,9 +1218,14 @@ class GeometryArray(ExtensionArray):
     def _pad_or_backfill(
         self, method, limit=None, limit_area=None, copy=True, **kwargs
     ):
-        return super()._pad_or_backfill(
-            method=method, limit=limit, limit_area=limit_area, copy=copy, **kwargs
-        )
+        if PANDAS_GE_21 and not PANDAS_GE_22:
+            if limit_area is not None:
+                # limit area not supported, but, but we feed through
+                # so the caller gets the pandas exception
+                kwargs["limit_area"] = limit_area
+        else:
+            kwargs["limit_area"] = limit_area
+        return super()._pad_or_backfill(method=method, limit=limit, copy=copy, **kwargs)
 
     def fillna(self, value=None, method=None, limit=None, copy=True):
         """
@@ -1640,7 +1643,7 @@ class GeometryArray(ExtensionArray):
         data = np.concatenate([ga._data for ga in to_concat])
         return GeometryArray(data, crs=_get_common_crs(to_concat))
 
-    def _reduce(self, name, skipna=True, **kwargs):
+    def _reduce(self, name, skipna=True, keepdims=False, **kwargs):
         # including the base class version here (that raises by default)
         # because this was not yet defined in pandas 0.23
         if name in ("any", "all"):
@@ -1672,7 +1675,7 @@ class GeometryArray(ExtensionArray):
                 ovalues = [param] * len(self)
             return ovalues
 
-        if isinstance(other, (pd.Series, pd.Index, pd.DataFrame)):
+        if isinstance(other, pd.Series | pd.Index | pd.DataFrame):
             # rely on pandas to unbox and dispatch to us
             return NotImplemented
 
