@@ -5,12 +5,12 @@ import os
 import pathlib
 import shutil
 import tempfile
+import warnings
 from collections import OrderedDict
 from packaging.version import Version
 
 import numpy as np
 import pandas as pd
-import pytz
 from pandas.api.types import is_datetime64_any_dtype
 
 from shapely.geometry import Point, Polygon, box, mapping
@@ -46,6 +46,7 @@ except ImportError:
     fiona = False
     FIONA_GE_19 = False
 
+pytz = pytest.importorskip("pytz")
 
 PYOGRIO_MARK = pytest.mark.skipif(not pyogrio, reason="pyogrio not installed")
 FIONA_MARK = pytest.mark.skipif(not fiona, reason="fiona not installed")
@@ -171,7 +172,7 @@ def test_to_file_pathlib(tmpdir, df_nybb, driver, ext, engine):
 @pytest.mark.parametrize("driver,ext", driver_ext_pairs)
 def test_to_file_bool(tmpdir, driver, ext, engine):
     """Test error raise when writing with a boolean column (GH #437)."""
-    tempfilename = os.path.join(str(tmpdir), "temp.{0}".format(ext))
+    tempfilename = os.path.join(str(tmpdir), f"temp.{ext}")
     df = GeoDataFrame(
         {
             "col": [True, False, True],
@@ -629,7 +630,7 @@ def test_read_file_geojson_string_path(engine):
             }
         ],
     }
-    df_read = read_file(json.dumps(features))
+    df_read = read_file(json.dumps(features), engine=engine)
     assert_geodataframe_equal(expected.set_crs("EPSG:4326"), df_read)
 
 
@@ -1075,6 +1076,31 @@ def test_read_file_mask_gdf_mismatched_crs(df_nybb, engine, nybb_filename):
     assert filtered_df_shape == (2, 5)
 
 
+@pytest.mark.skipif(not HAS_PYPROJ, reason="pyproj not installed")
+def test_read_file_multi_layer_with_layer_arg_no_warning(tmp_path, engine):
+    # While reading a file with multiple layers, if the layer is properly
+    # specified, a "Specify layer" warning should not be emitted
+    data1 = {"geometry": [Point(1, 2), Point(2, 1)]}
+    gdf = geopandas.GeoDataFrame(data1, crs="EPSG:4326")
+    file_path = tmp_path / "out.gpkg"
+    gdf.to_file(file_path, layer="layer1", engine=engine)
+    gdf.to_file(file_path, layer="layer2", engine=engine)
+
+    with warnings.catch_warnings(record=True) as captured:
+        warnings.simplefilter("always")
+        read_file(file_path, bbox=gdf, layer="layer1", engine=engine)
+        read_file(file_path, mask=gdf, layer="layer1", engine=engine)
+        specify_layer_warnings = [
+            warning
+            for warning in captured
+            if warning.category is UserWarning
+            and "specify layer parameter" in str(warning.message).lower()
+        ]
+        assert (
+            len(specify_layer_warnings) == 0
+        ), "'Specify layer parameter' warning was raised, but the layer was specified."
+
+
 def test_read_file_bbox_mask_not_allowed(engine, nybb_filename):
     bbox = (
         1031051.7879884212,
@@ -1119,7 +1145,7 @@ def test_read_file_empty_shapefile(tmpdir, engine):
     assert all(empty.columns == ["A", "Z", "geometry"])
 
 
-class FileNumber(object):
+class FileNumber:
     def __init__(self, tmpdir, base, ext):
         self.tmpdir = str(tmpdir)
         self.base = base
@@ -1127,7 +1153,7 @@ class FileNumber(object):
         self.fileno = 0
 
     def __repr__(self):
-        filename = "{0}{1:02d}.{2}".format(self.base, self.fileno, self.ext)
+        filename = f"{self.base}{self.fileno:02d}.{self.ext}"
         return os.path.join(self.tmpdir, filename)
 
     def __next__(self):
@@ -1399,7 +1425,6 @@ def test_option_io_engine(nybb_filename):
 
 @pytest.mark.skipif(pyogrio, reason="test for pyogrio not installed")
 def test_error_engine_unavailable_pyogrio(tmp_path, df_points, file_path):
-
     with pytest.raises(ImportError, match="the 'read_file' function requires"):
         geopandas.read_file(file_path, engine="pyogrio")
 
@@ -1409,7 +1434,6 @@ def test_error_engine_unavailable_pyogrio(tmp_path, df_points, file_path):
 
 @pytest.mark.skipif(fiona, reason="test for fiona not installed")
 def test_error_engine_unavailable_fiona(tmp_path, df_points, file_path):
-
     with pytest.raises(ImportError, match="the 'read_file' function requires"):
         geopandas.read_file(file_path, engine="fiona")
 
