@@ -542,8 +542,8 @@ class TestShapelyInterface:
                 pytest.xfail(
                     "rtree results are known to be unordered, see "
                     "https://github.com/geopandas/geopandas/issues/1337\n"
-                    "Expected:\n {}\n".format(expected)
-                    + "Got:\n {}\n".format(res.tolist())
+                    f"Expected:\n {expected}\n"
+                    f"Got:\n {res.tolist()}\n"
                 )
             raise e
 
@@ -767,8 +767,8 @@ class TestShapelyInterface:
                 pytest.xfail(
                     "rtree results are known to be unordered, see "
                     "https://github.com/geopandas/geopandas/issues/1337\n"
-                    "Expected:\n {}\n".format(expected)
-                    + "Got:\n {}\n".format(res.tolist())
+                    f"Expected:\n {expected}\n"
+                    f"Got:\n {res.tolist()}\n"
                 )
             raise e
 
